move copytree to compat
